requirements.txt
opencv-python
opencv-contrib-python
dlib-bin
scipy
numpy
